Fix checkpoint wal replay failed issue
Stop the cluster with immediate mode, the instance does not have
chance to write a SHUTDOWN_CHECKPOINT WAL. Hence, WAL replay is
trigger in startup process in next time cluster start. It will
collect a bunch of global txn which is not finished and write a
SHUTDOWN_CHECKPOINT to promise the cluster consistency.

Before we believe a RecordShort is enough for our extend checkpoint
wal due to global txn information. It is not always true, in data
bulk import into env, at least ten import session is working, so
RecordLong may used.

For every time the cluster instance start, it will valid the last
checkpoint wal size and replay. If the checkpoint wal was used
RecordLong the size validation believe it is an illegal xlog. The
issue is triggered.

Fix it.